DataViews: Expose FiltersToggled subcomponent.
